[TRTLLM-14554][refactor] clarify temporal batching configuration
