+++ tests/integration/test_lists/test-db/l0_b200.yml
@@ -221,7 +221,6 @@ l0_b200:
   - unittest/_torch/visual_gen/test_wan22_ti2v_5b_pipeline.py
   - unittest/_torch/visual_gen/test_wan_vsa_pipeline.py
   - unittest/_torch/visual_gen/test_wan_vae.py
-  - unittest/_torch/visual_gen/test_wan_vae_temporal_batching.py
   - unittest/_torch/visual_gen/test_wan21_i2v_teacache.py
   - unittest/_torch/visual_gen/test_wan21_t2v_teacache.py
   - unittest/_torch/visual_gen/test_wan21_t2v_teacache_user_coefficients.py
